+++ api/Controllers/StockController.cs
@@ -17,15 +17,19 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid) return BadRequest();
+
             var stocks = await _stockRepository.GetAllAsync();
             var stockDto = stocks.Select(s => s.ToStockDto());
 
             return Ok(stockDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest();
+
             var stock = await _stockRepository.GetByIdAsync(id);
             if (stock == null) return NotFound();
 
@@ -35,6 +39,8 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid) return BadRequest();
+
             var stockModel = stockDto.ToStockFromCreate();
             await _stockRepository.CreateAsync(stockModel);
 
@@ -42,9 +48,11 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid) return BadRequest();
+
             var stockModel = await _stockRepository.UpdateAsync(id, stockDto);
             if (stockModel == null) return NotFound();
 
@@ -52,9 +60,11 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest();
+
             await _stockRepository.DeleteAsync(id);
 
             return NoContent();
